requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cbor==1.0.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorlog==4.2.1
idna==2.10
lmdb==0.98
multidict==4.7.6
protobuf==3.12.2
pycparser==2.20
pyformance==0.4
pyperf==2.0.0
pyperformance==1.0.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
sawtooth-sdk==1.2.3
sawtooth-signing==1.1.5
secp256k1==0.13.2
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.2
urllib3==1.25.10
yarl==1.5.0
